\name{BiocParallelParam-class}
\Rdversion{1.1}
\docType{class}

\alias{BiocParallelParam}
\alias{BiocParallelParam-class}
\alias{show,BiocParallelParam-method}

\title{Virtual classes (for developer reference)}

\description{

  These classes are primarily relevant to developers.

  \code{BiocParallelParam} is a virtual class on which all parameter
  classes extend. It has no slots.

}

\author{Martin Morgan \url{mailto:mtmorgan@fhcrc.org}}

\examples{
getClass("BiocParallelParam")
}