\name{BiocParallel-defunct}

\alias{bprunMPIslave}
\alias{BatchJobsParam}

\title{Defunct Objects in Package \sQuote{BiocParallel}}

\description{

  These functions and objects are defunct and no longer available.

}

\details{

  Defunct functions are: \code{bprunMPIslave()}.

  Defunct classes: \code{BatchJobsParam}.

}
