\name{BiocParallel-deprecated}
\alias{BiocParallel-deprecated}
\title{Deprecated functions in package \sQuote{BiocParallel}}

\description{
  These functions are provided for compatibility with older versions
  of \sQuote{BiocParallel} only, and will be defunct at the next release.
}

\details{

  The following functions are deprecated and will be made defunct; use
  the replacement indicated below:
  \itemize{

    \item{BatchJobsParam: \code{\link{BatchtoolsParam}}}

  }
}
