\name{BiocParallel-package}
\alias{BiocParallel-package}
\alias{BiocParallel}
\docType{package}
\title{Bioconductor facilities for parallel evaluation}
\description{

  This package provides modified versions and novel implementation of
  functions for parallel evaluation, tailored to use with Bioconductor
  objects.

}

\details{

  This package uses code from the \code{\link{parallel}} package, 

}

\author{
  See \code{packageDescription("BiocParallel")}.
}

\keyword{package}
